webapp/requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
dlib-bin
opencv-python
numpy
face_recognition
Pillow
